+++ src/main/java/io/github/pod4dev/java/core/Utils.java
@@ -7,12 +7,10 @@ import java.io.BufferedReader;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.net.ServerSocket;
 import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
-import java.util.Random;
 
 @UtilityClass
 public final class Utils {
@@ -42,34 +40,6 @@ public final class Utils {
         }
     }
 
-    public static int findFreePort(List<Integer> binded) throws PodmanException {
-        Integer result = null;
-        final Random randomizer = new Random();
-        final int min = 30000;
-        final int max = 50000;
-        int counter = max - min;
-        while (result == null && counter > 0) {
-            int port = randomizer.nextInt(min, max);
-            if (binded.contains(port)) {
-                counter--;
-                continue;
-            }
-            try (ServerSocket serverSocket = new ServerSocket(port)) {
-                if (serverSocket.getLocalPort() == port) {
-                    result = port;
-                    break;
-                }
-            } catch (IOException ignored) {
-                counter--;
-                continue;
-            }
-        }
-        if (result == null) {
-            throw new PodmanException("There is no free port");
-        }
-        return result;
-    }
-
     public static URI getPodmanUri() {
         String podmanUri = System.getenv(Constants.ENV_PODMAN_HOST);
         if (podmanUri == null) {
@@ -79,10 +49,6 @@ public final class Utils {
             throw new PodmanException("No environment variable defined");
         }
         return URI.create(podmanUri);
-    }
-
-    public static String getUnixDomainSocket(URI podmanUri) {
-        return "unix".toLowerCase().equals(podmanUri.getScheme()) ? podmanUri.getPath() : null;
     }
 
     public static String getHost(URI podmanUri) {
